Move landing page to its own folder
