+++ src/components/Button.tsx
@@ -20,7 +20,7 @@ export default function Button({
     'px-7 py-3.5 font-semibold rounded-full transition-colors disabled:opacity-50 disabled:cursor-not-allowed';
 
   const variantStyles = {
-    primary: 'bg-orange-500 text-white hover:bg-orange-600',
+    primary: 'bg-primary-500 text-white hover:bg-primary-600',
     secondary: 'bg-black text-white hover:bg-gray-800',
     tertiary: 'bg-gray-100 text-black border border-gray-200 hover:bg-gray-200',
   };
