+++ src/components/icons/Logo.tsx
@@ -1,0 +1,88 @@
+import React from 'react';
+
+const Icon = ({ className = '' }: { className?: string }) => (
+  <svg
+    width='60'
+    height='47'
+    viewBox='0 0 60 47'
+    fill='none'
+    xmlns='http://www.w3.org/2000/svg'
+    className={className}
+  >
+    <path
+      d='M33.9517 5.13371C34.668 6.47128 34.668 6.47128 35.2428 8.2399L35.9072 10.1745C36.242 11.1886 36.5724 12.2051 36.9006 13.2223C38.5391 18.2319 40.0828 21.2594 44.422 24.0117C46.0546 25.1092 47.0739 25.8591 47.7813 27.803C48.1124 29.4049 48.3369 30.9833 48.5198 32.61L48.7517 34.2953C48.9369 35.6593 49.1057 37.0234 49.2606 38.3914L53.3346 39.9503H52.2235H7.77908L6.66797 39.9502L10.742 38.9113L10.9235 36.8566C11.3657 31.9623 11.8398 27.1156 12.7176 22.2821L13.962 22.1238C15.6443 21.8175 17.0828 21.4589 18.6435 20.7231C19.6509 19.1338 19.8272 17.3917 20.125 15.5272C20.588 12.8107 21.1154 10.5557 22.5939 8.25159C25.5198 6.11974 30.4361 3.69559 33.9517 5.13371ZM23.5813 9.29141C21.668 12.5901 20.9094 16.468 20.125 20.204L26.0509 18.6451C25.5569 19.6841 25.5569 19.6841 24.4732 20.0785L23.0583 20.4534L21.5243 20.8907C20.4517 21.1892 19.3761 21.4792 18.2991 21.7598C16.8517 22.1807 15.5383 22.6063 14.1991 23.3219C13.2657 24.9689 13.2657 24.9689 13.2109 26.9589C13.108 27.6854 13.0035 28.4118 12.8983 29.1375C12.8146 29.8936 12.7339 30.6512 12.6554 31.4089L12.5265 32.5563C12.2939 34.7154 12.2235 36.7132 12.2235 38.9113H23.5813L24.0443 35.2408C24.502 32.3481 25.0657 30.6544 27.3776 28.8101C28.0719 28.3394 28.7823 27.8953 29.5072 27.4788L30.385 27.9434C31.3317 28.4391 32.2865 28.9185 33.242 29.3947C35.2028 30.8383 35.4998 32.8236 35.8961 35.1761L36.4213 38.9113H47.7791L47.1924 33.6171L47.0265 31.9623L46.8628 30.6395L46.7139 29.295C45.8154 26.4967 43.345 25.018 41.1487 23.3725C37.8591 20.5696 36.6806 17.2708 35.3413 13.156C35.1057 12.4755 34.8687 11.7958 34.6302 11.1169C34.0598 9.4738 33.505 7.82756 32.9643 6.17352C29.7072 6.17352 26.105 7.02315 23.5813 9.29141ZM27.5324 30.1742L26.4517 31.1064C24.5124 33.3802 25.0628 35.9002 25.0628 38.9113H34.9398C35.1532 34.5938 35.1532 34.5938 33.4406 30.8975C30.465 28.3222 30.465 28.3222 27.5324 30.1742Z'
+      fill='currentColor'
+    />
+    <path
+      d='M27.8795 4.79406C29.8344 4.12834 31.9385 3.83356 33.7845 4.44982L34.1501 4.58522L34.3357 4.66095L34.4334 4.8434C34.7871 5.50401 34.9821 5.86708 35.1571 6.28805C35.3294 6.70257 35.4815 7.17106 35.7636 8.03907L36.4277 9.97369L36.4299 9.98058C36.7651 10.9958 37.0956 12.0132 37.4238 13.0305L37.7287 13.9393C38.438 16.004 39.1514 17.6763 40.1039 19.1281C41.1181 20.674 42.4192 21.9928 44.3183 23.2556L44.7068 23.508L44.7187 23.516C45.5299 24.0614 46.2408 24.5537 46.8357 25.1707C47.3676 25.7223 47.7971 26.362 48.1508 27.2131L48.2984 27.5918L48.3136 27.6331L48.3222 27.6779C48.6582 29.303 48.8846 30.9004 49.0688 32.5374L49.2999 34.2104C49.4694 35.4591 49.6236 36.709 49.7676 37.9603L53.5219 39.398C53.7756 39.4951 53.9266 39.7711 53.88 40.0521C53.8331 40.3328 53.6027 40.5375 53.3331 40.5375H6.66643C6.38498 40.5374 6.14842 40.3144 6.11521 40.0188C6.08221 39.7234 6.26274 39.4483 6.53622 39.3785L10.2233 38.4376L10.3687 36.8013C10.8109 31.9072 11.2878 27.0347 12.171 22.1712L12.2469 21.7501L13.8811 21.5412C15.4805 21.2494 16.826 20.9121 18.2658 20.2526C19.1145 18.8317 19.2791 17.2845 19.5755 15.4287L19.5766 15.423C20.043 12.6861 20.5872 10.3311 22.133 7.92203L22.1916 7.83024L22.2773 7.76713C23.7731 6.6773 25.782 5.50837 27.8795 4.79406ZM33.5675 5.61334C32.0156 5.05112 30.1283 5.26304 28.2202 5.91283C26.2885 6.57068 24.4137 7.64994 22.9935 8.67362C21.8337 10.5127 21.2962 12.3265 20.883 14.458C21.4223 12.5581 22.1128 10.6963 23.1063 8.98343L23.153 8.90426L23.2202 8.84344C25.8829 6.45042 29.6357 5.5858 32.9631 5.5858H33.358L33.4883 5.98168C34.0276 7.63139 34.5806 9.27346 35.1495 10.9123L35.8613 12.9536L35.8656 12.9651C37.2031 17.0743 38.3395 20.2219 41.4874 22.9079C42.5482 23.7015 43.7555 24.5135 44.7817 25.4587C45.8292 26.4236 46.7466 27.5744 47.2383 29.1053L47.2578 29.165L47.2643 29.2269L47.4119 30.5626L47.5757 31.8856L47.5779 31.9005L47.7417 33.5483L48.3873 39.3625H50.1647L49.0699 38.9436L48.7476 38.8209L48.7074 38.4606C48.5532 37.0976 48.3851 35.7388 48.2007 34.3802L47.9685 32.6946L47.9663 32.6796C47.7866 31.0811 47.5675 29.5379 47.2469 27.9774C46.9225 27.0985 46.5376 26.5072 46.0577 26.0095C45.558 25.4912 44.9432 25.0595 44.1219 24.5074C41.9021 23.0974 40.3724 21.6 39.1903 19.7983C38.0127 18.0033 37.197 15.9317 36.3735 13.4138V13.4115C36.0464 12.3976 35.7173 11.3851 35.3839 10.3753L34.7187 8.44069L34.7155 8.43036C34.4248 7.53615 34.2902 7.12126 34.1404 6.7608C34.0073 6.44064 33.8602 6.1609 33.5675 5.61334ZM35.5803 36.9367C35.5645 37.5039 35.5322 38.1573 35.4935 38.9413L35.4718 39.3625H35.921L35.5803 36.9367ZM24.1306 38.9884L24.084 39.3625H24.5061V38.9115C24.5061 38.1714 24.4724 37.4688 24.4507 36.76C24.4481 36.6743 24.4464 36.5886 24.4442 36.503L24.1306 38.9884ZM12.6929 25.8855C12.1622 29.5351 11.8106 33.2033 11.4755 36.9126L11.2932 38.9654L11.2584 39.3625H11.6664V38.9115C11.6664 36.6992 11.7378 34.6772 11.9735 32.4892V32.4869L12.1026 31.3394C12.1811 30.5823 12.2609 29.8246 12.3446 29.0686L12.3457 29.0583L12.3478 29.0479C12.4511 28.3354 12.5527 27.6221 12.6538 26.9091C12.6648 26.5108 12.6756 26.1778 12.6929 25.8855ZM32.5714 6.76654C29.5271 6.83198 26.2929 7.66631 24.0091 9.67765C22.338 12.5957 21.579 15.9837 20.8602 19.4047L27.0896 17.7661L26.546 18.909C26.3205 19.3831 26.1453 19.7744 25.8439 20.0484C25.5464 20.3187 25.1485 20.4531 24.6525 20.6336L24.6298 20.6416L24.607 20.6485L23.2007 21.0191L23.2018 21.0203L21.6675 21.4575L21.6643 21.4586C20.5934 21.7567 19.5191 22.0447 18.4438 22.325C17.0559 22.7286 15.8282 23.1301 14.5885 23.7742C14.1701 24.5143 14.001 24.8326 13.9082 25.1741C13.8086 25.5408 13.7932 25.9502 13.7649 26.9756L13.7639 27.0112L13.7584 27.0456C13.6554 27.7727 13.5512 28.4998 13.4459 29.2258C13.3632 29.9734 13.2838 30.7229 13.2061 31.4725L13.2051 31.4783L13.077 32.6257C12.8685 34.5623 12.7951 36.3702 12.783 38.324H23.0933L23.4915 35.1628L23.4937 35.1536L23.4948 35.1432C23.7255 33.6851 23.992 32.4646 24.5191 31.3658C25.0523 30.2543 25.8339 29.3047 27.0419 28.3411L27.0581 28.3273L27.0766 28.3159C27.7822 27.8375 28.5035 27.3852 29.2402 26.9618L29.4952 26.8161L29.7545 26.9527L30.6302 27.4162L32.0484 28.146L33.4774 28.8632L33.5197 28.8838L33.5577 28.9125C35.4623 30.3148 35.9436 32.2069 36.2953 34.2092L36.4418 35.0732L36.444 35.0893L36.8986 38.324H47.1525L46.6393 33.6848L46.6382 33.6791L46.4722 32.0244L46.3105 30.7152L46.3094 30.7072L46.1673 29.4289C45.7561 28.1884 44.9985 27.2168 44.0514 26.3445C43.0749 25.4451 41.9563 24.6994 40.8255 23.8522L40.8125 23.8408L40.7984 23.8293C37.3737 20.911 36.154 17.4637 34.8142 13.3472C34.5803 12.6715 34.3447 11.995 34.1078 11.3208L34.1067 11.3197C33.5814 9.80623 33.0715 8.28891 32.5714 6.76654ZM30.4045 29.4587C30.1762 29.4309 29.9157 29.4809 29.5028 29.679C29.0851 29.8795 28.5743 30.1985 27.847 30.6578L26.8389 31.5287C25.9884 32.5389 25.6788 33.5963 25.5824 34.7829C25.533 35.3916 25.5396 36.031 25.5607 36.7221C25.5762 37.2279 25.5996 37.7669 25.6107 38.324H34.4095C34.4967 36.5151 34.5217 35.6171 34.3661 34.7852C34.1995 33.8951 33.8245 33.066 32.9935 31.2717C32.2901 30.6631 31.7808 30.2252 31.3604 29.9234C30.9239 29.6102 30.642 29.4878 30.4045 29.4587ZM34.314 31.4565C34.4369 31.7242 34.5506 31.9721 34.6536 32.2057C34.5565 31.9441 34.4442 31.6951 34.314 31.4565ZM29.518 28.1414C28.8974 28.5053 28.2882 28.8901 27.6907 29.2946C27.5361 29.4185 27.3911 29.5422 27.2534 29.6641C27.9717 29.2105 28.5505 28.8447 29.0427 28.6085C29.3198 28.4755 29.5887 28.3755 29.8587 28.3216L29.518 28.1414ZM14.0547 22.7025L14.0265 22.7071L13.1866 22.8138C13.0714 23.4658 12.9651 24.1185 12.8633 24.7713C13.0165 24.2567 13.2901 23.7786 13.7205 23.0192L13.804 22.8734L13.9472 22.7977C14.0248 22.7563 14.1032 22.718 14.1805 22.6784C14.1387 22.6861 14.0968 22.6948 14.0547 22.7025ZM19.5462 20.2366C19.4378 20.4663 19.3174 20.6936 19.1816 20.9182C19.26 20.8973 19.3387 20.8772 19.4171 20.8562L19.5462 20.2366Z'
+      fill='currentColor'
+    />
+    <path
+      d='M27.714 4.27213C29.5923 3.63254 31.6711 3.30162 33.5715 3.81291L33.9487 3.92664L33.9653 3.93308L34.331 4.06827L34.3383 4.07149L34.3466 4.07471L34.5326 4.14981L34.7093 4.22277L34.8994 4.57792C35.2494 5.23144 35.4589 5.62021 35.6465 6.07147C35.8288 6.51003 35.9883 7.00188 36.2658 7.85578L36.9298 9.79031L36.935 9.80319L36.9371 9.8107C37.2713 10.823 37.6004 11.8369 37.9274 12.8504L38.2308 13.7559L38.4947 14.5038C39.1114 16.2044 39.7437 17.5991 40.5449 18.8203C41.5137 20.2968 42.7591 21.5643 44.6015 22.7902L44.9912 23.0434L44.9984 23.0477L45.0088 23.0552C45.8158 23.5977 46.5738 24.1189 47.2138 24.7827C47.7945 25.3849 48.2611 26.0834 48.6405 26.9961L48.6457 27.008L48.7932 27.3867L48.7964 27.3964L48.8275 27.48L48.8358 27.5251L48.8452 27.5702C49.184 29.2108 49.4117 30.8195 49.5965 32.4596L49.8272 34.1323V34.1345C49.9822 35.2763 50.1231 36.4188 50.2563 37.5615L53.7082 38.8834C54.178 39.0631 54.4574 39.5453 54.4179 40.0443L54.4054 40.1452C54.3162 40.6799 53.8718 41.0871 53.3341 41.0872H6.66706C6.13984 41.0871 5.70799 40.6969 5.60406 40.1859L5.58743 40.0819C5.52493 39.5224 5.86618 38.9833 6.40936 38.8447L9.72825 37.9971L9.83943 36.7514C10.2816 31.8574 10.7603 26.9612 11.6485 22.0702L11.7961 21.2537L13.8015 20.9973C15.3163 20.7205 16.5721 20.4042 17.8956 19.817C18.6031 18.551 18.7582 17.1733 19.0511 15.3396L19.0531 15.331L19.0542 15.3235C19.5236 12.5699 20.0829 10.1238 21.6904 7.61866L21.8026 7.4427L21.8836 7.38261L21.9689 7.31931L21.9709 7.31824C23.4982 6.20551 25.5548 5.00747 27.714 4.27213ZM33.7855 4.45024C31.9395 3.83399 29.8352 4.12788 27.8803 4.79358C25.783 5.50784 23.7743 6.67699 22.2785 7.76673L22.1923 7.83003L22.1341 7.92231L21.8556 8.37295C20.5203 10.6268 20.0152 12.8574 19.5779 15.4233L19.5768 15.4287L19.3659 16.7656C19.1548 18.0502 18.9031 19.187 18.2665 20.2527C16.8268 20.912 15.4809 21.2495 13.8815 21.5413L12.2481 21.7505L12.1722 22.1711L12.0112 23.0831C11.2294 27.6436 10.7839 32.2127 10.3694 36.8008L10.2239 38.4381L6.53717 39.378L6.43949 39.4134C6.22193 39.5167 6.08752 39.7599 6.11633 40.0186C6.14539 40.2772 6.32989 40.4804 6.56418 40.5271L6.66706 40.5379H53.3341L53.4328 40.5282C53.6583 40.4852 53.8397 40.2975 53.8807 40.0518C53.9214 39.8059 53.8109 39.564 53.6126 39.4424L53.5232 39.3984L49.7679 37.9606C49.624 36.7095 49.4698 35.4593 49.3003 34.2107L49.0696 32.5369C48.8855 30.9 48.6595 29.3025 48.3236 27.6775L48.3142 27.6335L48.2997 27.5916L48.1521 27.2129C47.8425 26.468 47.4745 25.8851 47.0309 25.3814L46.8366 25.1711C46.3905 24.7084 45.8791 24.3153 45.3091 23.9179L44.72 23.5155L44.7075 23.508L44.3189 23.2558C42.5387 22.0722 41.2841 20.8388 40.2986 19.4158L40.1043 19.1282C39.1518 17.6764 38.4392 16.0039 37.7299 13.9394L37.4244 13.0306L36.4311 9.98023L36.429 9.97379L35.765 8.03926C35.5534 7.38828 35.4147 6.9617 35.286 6.61438L35.1582 6.2882C34.9831 5.86723 34.7877 5.50355 34.4339 4.84294L34.3362 4.66054L34.1513 4.58543L33.7855 4.45024ZM12.6544 26.9092C12.5533 27.6221 12.4521 28.3352 12.3489 29.0476L12.3468 29.0584L12.3457 29.068C12.262 29.824 12.1821 30.5824 12.1036 31.3395L11.9748 32.4864V32.4897C11.7391 34.6774 11.6672 36.6992 11.6672 38.9113V39.3619H11.2588L11.476 36.9124C11.8112 33.2031 12.1632 29.5351 12.6938 25.8856C12.6765 26.1779 12.6653 26.5111 12.6544 26.9092ZM24.4513 36.76C24.473 37.4687 24.5074 38.1713 24.5074 38.9113V39.3619H24.0845L24.1312 38.9885L24.445 36.5025C24.4472 36.5881 24.4486 36.6743 24.4513 36.76ZM35.9219 39.3619H35.473L35.4938 38.9413C35.5325 38.1575 35.5653 37.5041 35.5811 36.937L35.9219 39.3619ZM28.2211 5.91267C30.1292 5.2629 32.0165 5.0511 33.5683 5.61332C33.8609 6.16056 34.0078 6.44036 34.1409 6.7603C34.2907 7.1207 34.426 7.53587 34.7165 8.42981L34.7197 8.44054L35.3847 10.3751C35.7181 11.3849 36.0468 12.3977 36.3739 13.4115V13.4137C37.1974 15.9314 38.0134 18.0029 39.1909 19.7977C40.3729 21.5993 41.9029 23.0969 44.1225 24.5069C44.9436 25.0589 45.5587 25.4909 46.0583 26.009C46.5382 26.5067 46.9237 27.0982 47.2481 27.9768C47.5686 29.5373 47.7874 31.0811 47.9672 32.6796L48.202 34.3802C48.3864 35.7387 48.5538 37.0977 48.708 38.4606L48.7486 38.8211L49.0707 38.9435L50.1659 39.3619H48.388L47.7427 33.5487L47.5785 31.9006L47.5765 31.8856L47.4123 30.5626L47.2647 29.2268L47.2585 29.1646L47.2388 29.1056C46.7471 27.5746 45.8298 26.4235 44.7823 25.4586C43.7563 24.5136 42.5491 23.7018 41.4884 22.9082C38.3405 20.2223 37.2042 17.0743 35.8668 12.9652L35.8616 12.9534L35.1499 10.9126C34.581 9.27376 34.0287 7.63105 33.4894 5.98134L33.3584 5.58542H32.9636C29.6363 5.58548 25.8835 6.45008 23.221 8.8429L23.1534 8.90406L23.1067 8.98345C22.1133 10.6962 21.4233 12.5579 20.884 14.4576C21.2972 12.3263 21.8347 10.5124 22.9944 8.67337C24.4146 7.64973 26.2895 6.57049 28.2211 5.91267ZM32.5718 6.76674C33.072 8.28897 33.5823 9.80597 34.1076 11.3193L34.1087 11.3203C34.3455 11.9945 34.5813 12.6714 34.8153 13.3471C36.1551 17.4636 37.3748 20.9106 40.7994 23.8288L40.8265 23.8524C41.9572 24.6995 43.0754 25.4456 44.0518 26.3449C44.9989 27.2171 45.7573 28.1881 46.1685 29.4285L46.3098 30.7075L46.3119 30.715L46.4729 32.024L46.6392 33.6796L46.6402 33.6849L47.1535 38.3244H36.8997L36.4446 35.0894L36.4425 35.0733L36.296 34.2096C35.9663 32.3325 35.5232 30.5513 33.8998 29.1807L33.559 28.9124L33.5205 28.8835L33.4779 28.8631L32.0492 28.1464L30.6308 27.4157L29.7559 26.9522L29.4961 26.8159L29.2415 26.9618C28.5048 27.3851 27.7827 27.8375 27.0771 28.3159L27.0594 28.3277L27.0428 28.3406L26.6095 28.7021C25.6434 29.5446 24.9865 30.3937 24.5198 31.3663L24.3338 31.7837C23.9266 32.7733 23.6972 33.8672 23.4953 35.1431L23.4943 35.1538L23.4922 35.1624L23.0942 38.3244H12.7842C12.7964 36.3706 12.8698 34.5625 13.0783 32.6259L13.2061 31.4779L13.2072 31.4725C13.2848 30.7228 13.3645 29.9734 13.4472 29.2257C13.5524 28.4998 13.6559 27.7726 13.7589 27.0455L13.7652 27.0112L13.7662 26.9758C13.7909 26.0787 13.8053 25.6528 13.8753 25.3148L13.9086 25.1743C14.0013 24.8328 14.1707 24.5141 14.5892 23.7741C15.8288 23.13 17.0565 22.7281 18.4442 22.3245C18.982 22.1844 19.52 22.0429 20.0569 21.8986L21.6654 21.4587L21.6686 21.4576L23.2023 21.0198L23.2012 21.0187L24.6082 20.6486L24.6539 20.6336C25.0877 20.4757 25.4462 20.3524 25.7283 20.1432L25.8447 20.0488C26.0707 19.8433 26.2255 19.5713 26.384 19.2484L26.5471 18.9093L27.0906 17.7656L20.8612 19.405C21.5799 15.9842 22.3386 12.5957 24.0096 9.67765C26.2934 7.6664 29.5277 6.83224 32.5718 6.76674ZM30.4053 29.4586C30.6428 29.4876 30.9248 29.6099 31.3613 29.9232C31.7817 30.2249 32.2914 30.6633 32.9948 31.2719C33.8256 33.0657 34.2009 33.8948 34.3674 34.7847C34.523 35.6166 34.4972 36.5155 34.41 38.3244H25.6119L25.5621 36.7224C25.5409 36.0313 25.5334 35.3912 25.5828 34.7826C25.6792 33.5961 25.9897 32.5383 26.8402 31.5283L27.8481 30.6581C28.5754 30.1988 29.0857 29.879 29.5034 29.6785C29.9163 29.4804 30.177 29.4307 30.4053 29.4586ZM32.1967 7.33004C29.3732 7.45276 26.4824 8.26108 24.4232 10.0371C22.9695 12.6093 22.2279 15.5658 21.5657 18.6518L28.0757 16.9394L27.0241 19.1507C26.8167 19.5869 26.5946 20.0983 26.1959 20.4608C25.8032 20.8175 25.287 20.9857 24.8305 21.1518L24.8253 21.1539L24.8025 21.1614L24.7911 21.1657L24.7568 21.1754L24.7484 21.1786L24.7391 21.1808L23.4319 21.5241L23.3446 21.5499L21.8275 21.9812L21.8203 21.9844L21.8036 21.9887C20.7298 22.2876 19.6527 22.5769 18.5752 22.8578L18.5741 22.8567C17.2535 23.241 16.1096 23.6178 14.9674 24.1947C14.6034 24.842 14.4868 25.0795 14.4208 25.3223C14.3432 25.6081 14.3262 25.9363 14.2972 26.9908V26.9929L14.2951 27.0637L14.2899 27.0992L14.2847 27.1335C14.182 27.8584 14.0779 28.5835 13.973 29.3073C13.8912 30.0467 13.813 30.7882 13.7361 31.5304L13.734 31.554L13.7319 31.5626L13.6062 32.6882C13.4201 34.417 13.3466 36.0442 13.3246 37.775H22.6266L22.9643 35.0916L22.9685 35.0615L22.9705 35.0497C23.2029 33.5825 23.4797 32.2959 24.0429 31.1216C24.6172 29.9247 25.4539 28.9159 26.7051 27.9157L26.7217 27.9028L26.7518 27.877L26.784 27.8567L26.8028 27.8449C27.5133 27.3638 28.2403 26.9088 28.9828 26.4822L28.9838 26.4811L29.4857 26.1936L29.9969 26.4629L29.998 26.464L30.8677 26.9243L32.2819 27.6517L33.7045 28.3663L33.7481 28.3867L33.7938 28.4092L33.8343 28.4403L33.8676 28.465L34.2406 28.759C35.7803 30.0609 36.3379 31.701 36.6763 33.3502L36.8197 34.1109L36.8207 34.1141L36.9662 34.9789L36.9683 34.9875L36.9693 34.9971L36.9714 35.01L37.36 37.775H46.5571L46.1113 33.7472L46.1103 33.7354L45.9451 32.0937L45.784 30.7901L45.783 30.7826L45.7819 30.7697L45.6458 29.5551C45.2703 28.4581 44.5902 27.5761 43.6985 26.7547C42.7487 25.8799 41.6763 25.1676 40.5137 24.2966L40.4971 24.2837L40.4815 24.2698L40.4784 24.2676L40.47 24.2612L40.4607 24.2526C36.9083 21.2255 35.6503 17.6392 34.3103 13.522C34.0906 12.8876 33.8695 12.2516 33.6473 11.6186L33.6463 11.6175L33.6068 11.5049C33.1248 10.1165 32.6568 8.72493 32.1967 7.33004ZM30.343 30.0047C30.2535 29.9938 30.098 29.9998 29.7278 30.1774C29.3492 30.3592 28.8748 30.6516 28.1557 31.1055L27.2174 31.9178C26.4778 32.8087 26.2025 33.7368 26.1138 34.8287C26.0671 35.4042 26.072 36.0173 26.093 36.7042C26.1033 37.0398 26.1147 37.4023 26.1263 37.775H33.9019C33.9677 36.3045 33.9715 35.5677 33.8447 34.8898C33.697 34.1003 33.3688 33.3561 32.5635 31.6163C31.9014 31.0437 31.4374 30.6473 31.0579 30.3749C30.6471 30.0801 30.4514 30.0181 30.343 30.0047ZM34.3144 31.4564C34.4445 31.6949 34.5571 31.9438 34.6542 32.2053C34.5512 31.9718 34.4373 31.724 34.3144 31.4564ZM29.8598 28.3212C29.5898 28.3752 29.3201 28.4758 29.0431 28.6088C28.551 28.845 27.9727 29.2112 27.2548 29.6646C27.3923 29.5427 27.5366 29.4182 27.6912 29.2944C28.2886 28.8899 28.8984 28.5049 29.519 28.141L29.8598 28.3212ZM13.948 22.7977L13.8046 22.8739L13.7215 23.0187C13.2913 23.7779 13.0176 24.2564 12.8643 24.7708C12.9661 24.1181 13.0722 23.4656 13.1874 22.8138L14.027 22.7065L14.0551 22.7022C14.0972 22.6945 14.14 22.6863 14.1818 22.6786C14.1045 22.7182 14.0256 22.7563 13.948 22.7977ZM19.4179 20.8557C19.3395 20.8766 19.2604 20.897 19.182 20.9179C19.3177 20.6934 19.4383 20.4662 19.5467 20.2366L19.4179 20.8557Z'
+      fill='currentColor'
+      strokeWidth={3}
+    />
+  </svg>
+);
+
+interface LogoProps {
+  variant?: 'iconOnly' | 'default' | 'withTagline';
+  colorScheme?: 'color' | 'white';
+}
+
+const Logo: React.FC<LogoProps> = ({
+  variant = 'default',
+  colorScheme = 'color',
+}) => {
+  const isWhite = colorScheme === 'white';
+
+  if (variant === 'iconOnly') {
+    return (
+      <Icon
+        className={`h-8 w-8 ${isWhite ? 'text-white' : 'text-primary-500'}`}
+      />
+    );
+  }
+
+  if (variant === 'withTagline') {
+    return (
+      <div className='flex flex-col'>
+        <div className='flex items-center gap-2'>
+          <Icon
+            className={`h-10 w-10 ${
+              isWhite ? 'text-white' : 'text-primary-500'
+            }`}
+          />
+          <span
+            className={`text-2xl font-semibold ${
+              isWhite ? 'text-white' : 'text-black'
+            }`}
+          >
+            Sonacove
+          </span>
+        </div>
+        <p className={`mt-2 ${isWhite ? 'text-white/80' : 'text-gray-600'}`}>
+          Teach Freely. Learn Joyfully.
+        </p>
+      </div>
+    );
+  }
+
+  // Default variant
+  return (
+    <a href='/' className='flex items-center gap-2'>
+      <Icon
+        className={`h-8 w-8 ${isWhite ? 'text-white' : 'text-primary-500'}`}
+      />
+      <span
+        className={`text-xl font-semibold ${
+          isWhite ? 'text-white' : 'text-black'
+        }`}
+      >
+        Sonacove
+      </span>
+    </a>
+  );
+};
+
+export default Logo;
