+++ src/pages/landing/components/CustomizationSection.tsx
@@ -1,8 +1,7 @@
 import React from 'react';
 
-// 1. Import your reusable component and the specific image for this feature
-import FeatureCard from '../pages/features/components/FeatureCard';
-import featureImage from '../assets/feature-customization.png';
+import FeatureCard from '../../features/components/FeatureCard';
+import featureImage from '../../../assets/feature-customization.png';
 
 // 2. Define the data for this specific feature as a single object
 const customizationFeatureData = {
