+++ src/pages/landing/components/FeatureSection.tsx
@@ -3,7 +3,7 @@ import feature_natural from '../../../assets/feature-natural.png';
 import feature_simple from '../../../assets/feature-simple.png';
 import feature_privacy from '../../../assets/feature-privacy.png';
 import feature_control from '../../../assets/feature-control.png';
-import FeatureCard from './FeatureCard';
+import FeatureCard from '../../features/components/FeatureCard';
 import SectionHeader from '../../../components/SectionHeader';
 
 const featureSections = [
