+++ src/pages/landing/components/Landing.tsx
@@ -1,0 +1,35 @@
+import React from 'react';
+
+import HeroSection from './HeroSection';
+import FeatureSection from './FeatureSection';
+import PricingSection from './PricingSection';
+import MetricsSection from './MetricsSection';
+import CTASection from './CTASection';
+import CertificationsSection from './CertificationsSection';
+import CustomizationSection from './CustomizationSection';
+
+const Landing: React.FC = () => {
+  return (
+      <main>
+        <HeroSection />
+
+        <div id='features'>
+          <FeatureSection />
+        </div>
+
+        <MetricsSection />
+
+        <CertificationsSection />
+
+        <CustomizationSection />
+
+        <div id='pricing'>
+          <PricingSection />
+        </div>
+
+        <CTASection />
+      </main>
+  );
+};
+
+export default Landing;
