+++ src/pages/meet.astro
@@ -5,5 +5,5 @@ import Layout from "../layouts/Layout.astro";
 ---
 
 <Layout title="Sonacove Meets">
-  <Meet client:only='react'/>
+  <Meet client:load/>
 </Layout>
